affile: handle EAGAIN properly
EAGAIN should not be reported as errors even in the case of the
file destination. One way to reproduce this problem is to tell syslog-ng
to log to stdout using file("/dev/stdout") and pressing Ctrl+S (XOFF)
to stop the output flow. Earlier, syslog-ng suspended the file destination
for time-reopen() seconds.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>